+++ uui-components/src/pickers/PickerToggler.tsx
@@ -4,6 +4,8 @@ import { IconContainer } from '../layout';
 import { i18n } from '../i18n';
 import css from './PickerToggler.module.scss';
 import { browserBugFixDirAuto } from '../helpers/browserBugFixDirAuto';
+import { flushSync } from 'react-dom';
+import { ControlIcon } from '../widgets/ControlIcon';
 
 export interface PickerTogglerRenderItemParams<TItem, TId> extends IHasCaption, IDisableable {
     /** DataRowProps object of the rendered item */
@@ -42,15 +44,16 @@ export interface PickerTogglerProps<TItem = any, TId = any>
     id?: string;
 }
 
-function PickerTogglerComponent<TItem, TId>(props: PickerTogglerProps<TItem, TId>, ref: React.ForwardedRef<HTMLElement>) {
+function PickerTogglerComponent<TItem, TId>(props: PickerTogglerProps<TItem, TId>, ref: React.ForwardedRef<HTMLDivElement>) {
     const [inFocus, setInFocus] = React.useState<boolean>(false);
 
-    const toggleContainer = React.useRef<HTMLDivElement>(undefined);
     const searchInput = React.useRef<HTMLInputElement>(undefined);
+    const containerRef = React.useRef<HTMLDivElement>(undefined);
 
-    React.useImperativeHandle(ref, () => toggleContainer.current, [toggleContainer.current]);
+    React.useImperativeHandle(ref, () => containerRef.current, [containerRef.current]);
 
     const isSearchInToggler = props.searchPosition === 'input';
+    const isPickerDisabled = props.isDisabled || props.isReadonly;
 
     const handleClick = React.useCallback(
         (event: Event) => {
@@ -91,29 +94,51 @@ function PickerTogglerComponent<TItem, TId>(props: PickerTogglerProps<TItem, TId
     };
 
     const handleBlur = (e: React.FocusEvent<HTMLElement>) => {
+        /*
+            Handles the case when focus is moved from the input
+            to the picker body when it opens any any interactions happen there.
+            It is necessary to exit because otherwise `isFocus` will be set to `true`
+            (which will trigger `handleClick`) and would close the picker body.
+        */
         if (props.isOpen) {
-            // If picker opened and search inside input, we lock focus on toggler.
-            // In case, when search inside body, we need to highlight toggler like in focus state, even when focus was moved to the body. So we do nothing in this case.
-            if (isSearchInToggler) {
-                searchInput.current?.focus();
-            }
-        } else {
-            // If picker closed, we perform blur event as usual.
-            blur(e);
+            return;
         }
+
+        blur(e);
+    };
+
+    const getFocusableControl = (): HTMLElement | undefined => {
+        if (searchInput.current) {
+            return searchInput.current;
+        }
+
+        return containerRef.current;
+    };
+
+    const moveFocusToControl = () => {
+        const focusableControl = getFocusableControl();
+
+        focusableControl?.focus();
+
+        setInFocus(true);
     };
 
     const handleCrossIconClick = () => {
         if (props.onClear) {
-            props.onClear();
-            props.onValueChange('');
+            /*
+                `flushSync` is necessary for case when the input was hidden, and after the action is appears again.
+                In this case, `moveFocusToControl` should pick it instead of the container.
+            */
+            flushSync(() => {
+                props.onClear();
+                props.onValueChange('');
+            });
         }
-        // When we click on the cross it disappears from the DOM and focus is passed to the Body. So in this case we have to return focus on the toggleContainer by hand.
-        toggleContainer.current?.focus();
+        // When we click on the cross it disappears from the DOM and focus is passed to the Body. So in this case we have to return focus on the control by hand.
+        moveFocusToControl();
     };
 
     const renderItems = () => {
-        const isPickerDisabled = props.isDisabled || props.isReadonly;
         let areAllDisabled = isPickerDisabled;
         const displayedRows = props.selectedRowsCount > props.maxItems ? props.selection.slice(0, props.maxItems) : props.selection;
         const collapsedRows = props.selection?.slice(props.maxItems);
@@ -129,9 +154,15 @@ function PickerTogglerComponent<TItem, TId>(props: PickerTogglerProps<TItem, TId
                 isCollapsed: false,
                 isDisabled: isPickerDisabled || row.isDisabled,
                 onClear: () => {
-                    row.onCheck?.(row);
-                    // When we delete item it disappears from the DOM and focus is passed to the Body. So in this case we have to return focus on the toggleContainer by hand.
-                    toggleContainer.current?.focus();
+                    /*
+                        `flushSync` is necessary for case when the input was hidden, and after the action is appears again.
+                        In this case, `moveFocusToControl` should pick it instead of the container.
+                    */
+                    flushSync(() => {
+                        row.onCheck?.(row);
+                    });
+                    // When we delete item it disappears from the DOM and focus is passed to the Body. So in this case we have to return focus on the control by hand.
+                    moveFocusToControl();
                 },
             };
 
@@ -154,6 +185,19 @@ function PickerTogglerComponent<TItem, TId>(props: PickerTogglerProps<TItem, TId
         return tags;
     };
 
+    const shouldToggleBody = (e: React.MouseEvent<HTMLDivElement>): boolean => {
+        const isInteractionDisabled = isPickerDisabled || isEventTargetInsideClickable(e);
+        const shouldOpenWithMinCharsToSearch = (inFocus && props.value && (props.minCharsToSearch && props.searchPosition === 'input'));
+        const isPickerOpenWithSearchInInput = (props.isOpen && props.searchPosition === 'input' && (e.target as HTMLInputElement).tagName === 'INPUT');
+        return !(isInteractionDisabled || shouldOpenWithMinCharsToSearch || isPickerOpenWithSearchInInput);
+    };
+
+    const togglerPickerOpened = (e: React.MouseEvent<HTMLDivElement>) => {
+        e.preventDefault();
+        if (!shouldToggleBody(e)) return;
+        props.onClick?.();
+    };
+
     const renderInput = () => {
         const isSinglePickerSelected = props.pickerMode === 'single' && props.selection && !!props.selection[0];
         let placeholder: string;
@@ -167,6 +211,8 @@ function PickerTogglerComponent<TItem, TId>(props: PickerTogglerProps<TItem, TId
         const value = props.disableSearch ? null : props.value;
 
         if (!isSearchInToggler && props.pickerMode === 'multi' && props.selectedRowsCount > 0) {
+            searchInput.current = undefined;
+
             return null;
         }
 
@@ -174,7 +220,7 @@ function PickerTogglerComponent<TItem, TId>(props: PickerTogglerProps<TItem, TId
             <input
                 id={ props?.id }
                 type="search"
-                tabIndex={ props.isReadonly || props.isDisabled || !isSearchInToggler ? -1 : 0 } // If search not in toggler, we shouldn't make this input focusable
+                tabIndex={ isPickerDisabled || !isSearchInToggler ? -1 : 0 } // If search not in toggler, we shouldn't make this input focusable
                 ref={ searchInput }
                 aria-haspopup={ true }
                 autoComplete="no"
@@ -194,53 +240,112 @@ function PickerTogglerComponent<TItem, TId>(props: PickerTogglerProps<TItem, TId
                 value={ value || '' }
                 readOnly={ props.isReadonly || props.disableSearch }
                 onChange={ (e) => props.onValueChange?.(e.target.value) }
+                onKeyDown={ props.onKeyDown }
                 dir={ browserBugFixDirAuto(value || placeholder) } // TODO: remove after browser bug fix
+                onClick={ togglerPickerOpened }
             />
         );
     };
 
-    const shouldToggleBody = (e: React.MouseEvent<HTMLDivElement>): boolean => {
-        const isInteractionDisabled = (props.isDisabled || props.isReadonly || isEventTargetInsideClickable(e));
-        const shouldOpenWithMinCharsToSearch = (inFocus && props.value && (props.minCharsToSearch && props.searchPosition === 'input'));
-        const isPickerOpenWithSearchInInput = (props.isOpen && props.searchPosition === 'input' && (e.target as HTMLInputElement).tagName === 'INPUT');
-        return !(isInteractionDisabled || shouldOpenWithMinCharsToSearch || isPickerOpenWithSearchInInput);
-    };
-
-    const togglerPickerOpened = (e: React.MouseEvent<HTMLDivElement>) => {
-        e.preventDefault();
-        if (!shouldToggleBody(e)) return;
-        toggleContainer.current.focus();
-        props.onClick?.();
-    };
-
-    const icon = props.icon && (
-        <IconContainer
+    const icon = (
+        <ControlIcon
             icon={ props.icon }
             onClick={ props.onIconClick }
+            isDisabled={ isPickerDisabled }
+            rawProps={ {
+                'aria-label': 'Icon in input',
+            } }
         />
     );
 
+    const getIsEventTargetContainer = (event: React.SyntheticEvent): boolean => {
+        const eventTargetElement = event.target as HTMLElement;
+
+        return eventTargetElement === containerRef.current;
+    };
+
+    const getIsNonClickableEventTarget = (event: React.SyntheticEvent): boolean => {
+        const eventTargetElement = event.target as HTMLElement;
+        const isWithingContainer = containerRef.current?.contains(eventTargetElement);
+        const isClickable = eventTargetElement.classList.contains(uuiMarkers.clickable);
+
+        return (
+            (
+                isWithingContainer
+                && !isClickable
+            )
+            || getIsEventTargetContainer(event)
+        );
+    };
+
+    const getIsFocusableControlContainer = (): boolean => {
+        return containerRef.current === getFocusableControl();
+    };
+
+    const handleWrapperClick = (event: React.MouseEvent<HTMLElement>) => {
+        if (isPickerDisabled) {
+            return;
+        }
+
+        if (getIsNonClickableEventTarget(event)) {
+            moveFocusToControl();
+
+            if (
+                getIsFocusableControlContainer()
+                && getIsFocusableControlContainer()
+            ) {
+                props.onClick?.(event);
+            }
+        }
+    };
+
+    const handleWrapperFocus = (event: React.FocusEvent<HTMLElement>) => {
+        if (isPickerDisabled) {
+            return;
+        }
+
+        if (getIsNonClickableEventTarget(event)) {
+            moveFocusToControl();
+
+            if (getIsEventTargetContainer(event)) {
+                props.onFocus?.(event);
+            }
+        }
+    };
+
+    const handleKeyDown = (event: React.KeyboardEvent<HTMLElement>) => {
+        if (isPickerDisabled) {
+            return;
+        }
+
+        if (
+            getIsNonClickableEventTarget(event)
+            && getIsFocusableControlContainer()
+        ) {
+            props.onKeyDown?.(event);
+        }
+    };
+
     return (
         <div
-            onClick={ togglerPickerOpened }
-            ref={ toggleContainer }
+            ref={ containerRef }
             className={ cx(
                 css.container,
                 uuiElement.inputBox,
                 props.isDisabled && uuiMod.disabled,
                 props.isReadonly && uuiMod.readonly,
                 props.isInvalid && uuiMod.invalid,
-                !props.isReadonly && !props.isDisabled && props.onClick && uuiMarkers.clickable,
                 !props.isReadonly && !props.isDisabled && inFocus && uuiMod.focus,
                 props.selection?.length > 0 && uuiMarkers.hasValue,
                 props.cx,
             ) }
-            onFocus={ handleFocus }
+            onClick={ handleWrapperClick }
+            onFocus={ handleWrapperFocus }
             onBlur={ handleBlur }
+            onKeyDown={ handleKeyDown }
             // If search in toggler, we make nested search input focusable, and wrapper not.
             // It's required that only 1 focusable element in toggler, since if we have more we will go through all of them using tab key
-            tabIndex={ props.isReadonly || props.isDisabled || isSearchInToggler ? undefined : 0 }
-            onKeyDown={ props.onKeyDown }
+            tabIndex={ isPickerDisabled || isSearchInToggler ? undefined : 0 }
             { ...props.rawProps }
         >
             <div className={ cx(css.body, !props.isSingleLine && props.pickerMode !== 'single' && 'uui-picker_toggler-multiline') }>
@@ -249,21 +354,33 @@ function PickerTogglerComponent<TItem, TId>(props: PickerTogglerProps<TItem, TId
                 {renderInput()}
                 {props.iconPosition === 'right' && icon}
             </div>
-            {!props.isDisabled && !props.isReadonly && (
+            {!isPickerDisabled && (
                 <div className="uui-picker_toggler-actions">
                     {!props.disableClear && (props.value || props.selectedRowsCount > 0) && (
-                        <IconContainer
-                            cx={ cx('uui-icon-cancel', uuiMarkers.clickable) }
+                        <ControlIcon
+                            cx="uui-icon-cancel"
                             isDisabled={ props.isDisabled }
                             icon={ props.cancelIcon }
-                            tabIndex={ -1 }
                             onClick={ handleCrossIconClick }
-                            rawProps={ { role: 'button', 'aria-label': 'Clear' } }
+                            rawProps={ {
+                                'aria-label': 'Clear input',
+                            } }
                         />
                     )}
                     {props.isDropdown
                         && (!props?.minCharsToSearch || (props?.minCharsToSearch && props.searchPosition === 'body'))
-                        && <IconContainer icon={ props.dropdownIcon } flipY={ props.isOpen } cx="uui-icon-dropdown" />}
+                        && (
+                            <IconContainer
+                                icon={ props.dropdownIcon }
+                                flipY={ props.isOpen }
+                                cx="uui-icon-dropdown"
+                                /*
+                                    All clicks on non-interactive elements switch focus to the toggler,
+                                    but in this case it is necessary to toggler the dropdown.
+                                */
+                                onClick={ togglerPickerOpened }
+                            />
+                        )}
                 </div>
             )}
         </div>