+++ changelog.md
@@ -1,3 +1,24 @@
+# 6.x.x - xx.xx.xxx
+
+* Make icons and clear buttons in input accessible via keyboard ([#2845](https://github.com/epam/UUI/pull/2845))
+    * Icon set with `icon` property and clear button set with `onClear` property now use a button component with its own CSS variables, which makes some input-specific ones obsolete.
+        * The following CSS variables are no longer available:
+            * `TextInput` (and other inputs that use it):
+                * `--uui-text_input-icon`
+                * `--uui-text_input-icon-disabled`
+                * `--uui-text_input-icon-clickable`
+                * `--uui-text_input-icon-clickable-hover`
+                * `--uui-text_input-icon-clickable-active`
+            * `PickerInput`:
+                * `--uui-picker_toggler-icon-disabled`
+                * `--uui-picker_toggler-icon-clickable`
+                * `--uui-picker_toggler-icon-clickable-hover`
+        * The following CSS variables are should be used instead:
+            * `--uui-control_icon`
+            * `--uui-control_icon-clickable`
+            * `--uui-control_icon-clickable-hover`
+            * `--uui-control_icon-clickable-active`
+
 # 6.3.0 - 26.09.2025
 
 **What's New**